+++ build-logic/logic/src/main/kotlin/io/github/ackeecz/guardian/plugin/TestingPlugin.kt
@@ -43,9 +43,7 @@ internal class TestingPlugin : Plugin<Project> {
 
             // Kotest
             testImplementation(libs.kotest.assertions.core)
-            testImplementation(libs.kotest.framework.api)
-            testImplementation(libs.kotest.framework.datatest)
-            testRuntimeOnly(libs.kotest.runner.junit5)
+            testImplementation(libs.kotest.runner.junit5)
         }
     }
 }