+++ core-internal/src/testFixtures/kotlin/io/github/ackeecz/guardian/core/internal/Utils.kt
@@ -6,24 +6,24 @@ import io.kotest.matchers.shouldBe
 import org.robolectric.shadow.api.Shadow
 import java.nio.ByteBuffer
 
-val String.utf8ByteSize: Int
+public val String.utf8ByteSize: Int
     get() = toByteArray(Charsets.UTF_8).size
 
 /**
  * Returns [ByteBuffer.remaining] number of bytes in this [ByteBuffer] as [ByteArray]
  */
-fun ByteBuffer.remainingToByteArray(): ByteArray {
+public fun ByteBuffer.remainingToByteArray(): ByteArray {
     return ByteArray(remaining()).also { get(it) }
 }
 
-infix fun ByteString.shouldHaveBitSize(bitSize: Int) {
+public infix fun ByteString.shouldHaveBitSize(bitSize: Int) {
     toByteArray()
         .size
         .let { it * 8 }
         .shouldBe(bitSize)
 }
 
-fun Context.getKeysetGetCallCount(prefsName: String, keysetAlias: String): Int {
+public fun Context.getKeysetGetCallCount(prefsName: String, keysetAlias: String): Int {
     val prefs = getSharedPreferences(prefsName, Context.MODE_PRIVATE)
     val shadowPrefs = Shadow.extract<ShadowSharedPreferences>(prefs)
     return shadowPrefs.getGetCallCount(keysetAlias)