+++ core-internal/src/testFixtures/kotlin/io/github/ackeecz/guardian/core/internal/kotest/extension/CoroutineExtension.kt
@@ -3,14 +3,14 @@ package io.github.ackeecz.guardian.core.internal.kotest.extension
 import io.kotest.core.listeners.AfterEachListener
 import io.kotest.core.listeners.BeforeEachListener
 import io.kotest.core.test.TestCase
-import io.kotest.core.test.TestResult
+import io.kotest.engine.test.TestResult
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.TestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
 
-class CoroutineExtension(
+public class CoroutineExtension(
     private val testDispatcher: TestDispatcher = StandardTestDispatcher(),
 ) : BeforeEachListener, AfterEachListener {
 
@@ -18,7 +18,7 @@ class CoroutineExtension(
         beforeEach()
     }
 
-    fun beforeEach() {
+    public fun beforeEach() {
         Dispatchers.setMain(testDispatcher)
     }
 
@@ -26,7 +26,7 @@ class CoroutineExtension(
         afterEach()
     }
 
-    fun afterEach() {
+    public fun afterEach() {
         Dispatchers.resetMain()
     }
 }